Moved bootstrap and inspectIT core to separated subpackages.
