--- inspectit-oce-agent/src/main/java/rocks/inspectit/oce/agent/AgentImpl.java
@@ -1,10 +1,0 @@
-package rocks.inspectit.oce.agent;
-
-import rocks.inspectit.oce.agent.bootstrap.IAgent;
-
-public class AgentImpl implements IAgent {
-
-    public static void createInstance() {
-        //TODO: Agent initializing code residing in the inspectIT classloader
-    }
-}
