--- inspectit-oce-agent/src/main/java/rocks/inspectit/oce/agent/bootstrap/IAgent.java
@@ -1,5 +1,0 @@
-package rocks.inspectit.oce.agent.bootstrap;
-
-public interface IAgent {
-
-}
