+++ settings.gradle
@@ -1,1 +1,4 @@
 include "inspectit-oce-agent"
+include 'inspectit-oce-bootstrap'
+include 'inspectit-oce-core'
+
